+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `ui/Scroller` scroll animation method for `scrollMode: native`
+
 ## [5.3.0] - 2025-09-24
 
 ### Added